chore: rebase 449-refactor
