+++ libs/modules/app/src/core/application/handlers/GetAppsQueryHandler.ts
@@ -3,7 +3,7 @@ import { IQueryHandler, QueryHandler } from '@nestjs/cqrs'
 import { fold } from 'fp-ts/lib/Either'
 import { AppDITokens } from '../../../framework/AppDITokens'
 import { App } from '../../domain/app'
-import { GetAppsQuery } from '../commands/GetAppsQuery'
+import { GetAppsQuery } from '../queries/GetAppsQuery'
 import { GetAppsUseCase } from '../useCases/getApps/GetAppsUseCase'
 import { Result } from '@codelab/backend'
