+++ libs/modules/app/src/core/application/queries/GetAppQuery.ts
