+++ libs/modules/app/src/core/application/useCases/getApp/GetAppService.ts
@@ -16,6 +16,6 @@ export class GetAppService {
       return left(new GetAppErrors.NotFound(appId))
     }
 
-    return right(Result.ok(app.value))
+    return right(Result.ok(app))
   }
 }
